% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicability_domain.R
\name{ApplicabilityDomain}
\alias{ApplicabilityDomain}
\title{Compute applicability domain for a chemmodlab model}
\usage{
ApplicabilityDomain(traindata, testdata, pvalue = 0.01, desname = NULL)
}
\description{
\code{ApplicabilityDomain} evaluates the applicability domain for a chemmodlab model
using a Hotteling T2 control chart.
}
