% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_curve.R
\name{HitCurve}
\alias{HitCurve}
\title{Plot hit enrichment curves from observed scores and activities}
\usage{
HitCurve(
  S.df,
  labels = NULL,
  y,
  y.bin = TRUE,
  x.max = NULL,
  log = TRUE,
  title = "",
  conf = FALSE,
  conf.level = 0.95,
  method = "sup-t",
  plus = TRUE,
  band.frac = NULL
)
}
\arguments{
\item{S.df}{Data frame where variables are numeric scores from different
algorithms. Rows represent unique compounds.}

\item{labels}{Character vector of labels for the different algorithms in
\code{S.df}. Variable names in \code{S.df} will be used by default.}

\item{y}{Numeric vector of non-negative activity values. Compounds are
assumed to in the same order as in \code{S.df}. The value 0 is assumed to
be undesirable. If \code{y.bin} is \code{TRUE}, then \code{y} will be
converted using \code{y <- as.integer(y != 0)}.}

\item{y.bin}{Logical. \code{TRUE} converts \code{y} to a vector or 0 and 1.}

\item{x.max}{Integer, the maximum number of tests allowed on the x axis.}

\item{log}{Logical. \code{TRUE} plots the x axis on the log scale.}

\item{title}{Character string}

\item{conf}{Logical. \code{TRUE} plots (simultaneous) confidence bands for
all hit enrichment curves.}

\item{conf.level}{Numeric, confidence coefficient}

\item{method}{Character indicates the method used to obtain confidence bands.
The default is \code{sup-t} but other options (not recommended) are
"theta-proj" and "bonf".}

\item{plus}{Logical. \code{TRUE} uses plus-adjusted version of \code{method}.}

\item{band.frac}{Numeric vector of fractions tested to be used in obtaining
confidence bands. Vector should be no longer than \code{y}, and should have
at least 20 entries. Entries should be in (0,1]. It is recommended that
entries be consistent with between 1 and \code{x.max} tests.}
}
\description{
Plot hit enrichment curves based on scores from multiple algorithms. Actual
activities are required. Additionally plot the ideal hit enrichment curve
that would result under perfect scoring, and the hit enrichment curve that
would result under random scoring. Optionally, simultaneous confidence bands
may also be requested.
}
