% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_assess.R
\name{CombineSplits}
\alias{CombineSplits}
\title{ANOVA and multiple comparisons for chemmodlab objects}
\usage{
CombineSplits(cml.result, metric = "enhancement", at = NA, thresh = 0.5,
  ...)
}
\arguments{
\item{cml.result}{an object of class \code{\link{chemmodlab}}.}

\item{metric}{the model performance measure to use.  This should be
one of \code{"error rate"}, \code{"enhancement"}, \code{"R2"},
\code{"rho"}, \code{"auc"}, \code{"sensitivity"}, \code{"specificity"}.}

\item{at}{the number of tests to use for initial \code{enchancement}.}

\item{thresh}{if the predicted probability that a binary response is 
1 is above this threshold, an observation is classified as 1. Used to
to compute \code{"error rate"}, \code{"sensitivity"}, and 
\code{"specificity"}}
}
\description{
\code{CombineSplits} assesses observed performance measures
across all splits created by \code{\link{ModelTrain}}, performs
statistical tests to determine the best performing descriptor set and
model (D-M) combinations.
}
\details{
\code{CombineSplits}
quantifies how sensitive performance measures are to fold
assignments (assignments to training and test sets). 
Intuitively, this
assesses how much a performance measure may change if a slightly
different data set is used.

\code{CombineSplits} is a designed study in that 'experimental' 
conditions are defined according to two factors: method (D-M combination) 
and split (fold assignment).  As such, an analysis
of variance to identify significant differences between 
performance measures according to factors and levels is appropriate.

The multiple comparisons plot shows the results for tests for signficance
in all pairwise differences of D-M mean performance measures.
Because there can be many 
estimated mean performance measures for an assay, care must be taken 
to adjust for
multiple testing, and we do this using the Tukey-Kramer multiple
comparison procedure (see Tukey (1953) and Kramer (1956)).
If you are having trouble viewing all the components of the plot,
make the plotting window larger.

By default, \code{CombineSplits} uses initial enhancement
proposed by Kearsley et al. (1996) to assess model performance. 
Enhancement at \code{n} tests is the hit
rate at \code{n} tests (accumulated actives at \code{n} tests
divided by \code{n}, the 
number of tests) divided by the proportion of actives in the entire 
collection. It is a relative measure of hit rate improvement offered
by the new method beyond what can be expected under random selection,
and values much larger than one are desired. Initial enhancement is
typically taken to be enhancement at \code{n}=300 tests.

Root mean squared error (\code{RMSE}), despite its popularity
in statistics, may be  
inappropriate for continuous chemical assay responses because
it assumes losses 
are equal for both under-predicting and over-predicting biological 
activity.  A suitable alternative may be initial \code{enhancement}.
Other options are Pearson's \code{R2} and Spearman's \code{rho}.

For binary chemical assay responses, alternatives to 
misclassification rate (\code{error rate}) 
(which may be inappropriate because it assigns equal weights to false
positives and false negatives) include \code{sensitivity},
\code{specificity},
area under the receiver operating characteristic curve (\code{ROC}),
and initial \code{enhancement}.
}
\examples{
# A data set with  binary response and multiple descriptor sets

cml <- ModelTrain(aid364, ids = T, xcol.lengths = c(24, 147), 
                  des.names = c("BurdenNumbers", "Pharmacophores"))
CombineSplits(cml)

# A continuous response

cml <- ModelTrain(USArrests)
CombineSplits(cml)
}
\author{
Jacqueline Hughes-Oliver, Jeremy Ash
}
\references{
Kearsley, S.K., Sallamack, S., Fluder, E.M., Andose, J.D., Mosley, R.T.,
and Sheridan, R.P. (1996). Chemical similarity using physiochemical
property descriptors, J. Chem. Inf. Comput. Sci. 36, 118-127.

Kramer, C. Y. (1956). Extension of multiple range tests to group means
with unequal numbers of replications. Biometrics 12, 307-310.

Tukey, J. W. (1953). The problem of multiple comparisons. Unpublished
manuscript. In The Collected Works of John W. Tukey VIII. Multiple
Comparisons: 1948-1983, Chapman and Hall, New York.
}
\seealso{
\code{\link{chemmodlab}}, \code{\link{ModelTrain}}
}

