% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_anova.R
\name{CombineSplits}
\alias{CombineSplits}
\title{Anova and multiple comparisons for chemmodlab objects}
\usage{
CombineSplits(cml.result, file = NA, metric = "enhancement", at = NA,
  thresh = 0.5, ...)
}
\description{
Anova and multiple comparisons for chemmodlab objects
}
\examples{
cml <- modelTrain(USArrests, nsplits = 3)
CombineSplits(cml)

bin <- rbinom(50, 1, .1)
cml <- ModelTrain(cbind(bin, USArrests), nsplits = 3)
CombineSplits(cml)
}

