% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{ModelTrain}
\alias{ModelTrain}
\title{Fit predictive models to sets of descriptors.}
\usage{
ModelTrain(data, ycol = ifelse(is.na(idcol), 1, 2),
  xcols = ifelse(is.na(idcol), list(seq(2, ncol(data))), list(seq(3,
  ncol(data)))), idcol = NA, nfolds = 10, nsplits = 1, seed.in = NA,
  des.names = NA, models = c("NNet", "PLS", "LARs", "PLSLDA", "Tree", "SVM",
  "KNN", "Forest"), user.params = NULL)
}
\arguments{
\item{data}{A data frame containing a response column, descriptor columns
and an (optional) id column.}

\item{ycol}{The index of the response column.}

\item{xcols}{A list of integer vectors.  Each vector contains
column indices
of \code{data} where a set of descriptor variables is located.
Users can specify multiple descriptor sets. By default there is one
descriptor set, all columns in \code{data} except \code{ycol} and
\code{idcol}.}

\item{idcol}{The index of the column of \code{data} containing
the compound labels. By default this is \code{NA}, and the row labels
of \code{data} will be used.}

\item{nfolds}{The number of folds to use for each cross
validation split.}

\item{nsplits}{The number of splits to use for repeated
cross validation.}

\item{seed.in}{A numeric vector with length equal to \code{nsplits}.
The seeds are used to randomly assign folds to observations for each
repeated cross-validation split.}

\item{des.names}{A character vector specifying the names for each
descriptor
set.  The length of the vector must match the number of descriptor sets.
If \code{NA}, descriptor set i will be named "Descriptor Set i".}

\item{models}{A character vector specifying the regression or
classification models to use.  The strings must match models
implemented in `chemmodlab` (see Details).}

\item{user.params}{A list of data frames where each data frame contains
the parameters values for a model.  The list should have the format of
the list constructed by \code{\link{MakeModelDefaults}}. One can construct
a list of parameters using \code{\link{MakeModelDefaults}} and then
modify the parameters.}
}
\value{
A list is returned of class \code{\link{chemmodlab}} containing:
 \item{all.preds}{a list of lists of data frames.  The elements of the outer
  list correspond to each split performed by ModelTrain. The
  elements of the inner list correspond to each descriptor set.  The elements
  of the inner list are data frames containing all model predictions for a
  split and descriptor set combination.  The first column of each data frame
  contains the true value of the response.  The remaining columns contain
  the predictions for each model.}
\item{all.probs}{a list of lists of data frames. Constructed only if there is
  a binary response.  The structure is the same as \code{all.preds}, except
  that predictions are replaced by predicted probabilities.  Predicted
  probabilities are only reported for classification models.}
\item{model.acc}{a list of lists of model accuracy measures.  The elements of
  the outer list correspond to each split performed by ModelTrain.
  The elements of the inner list correspond to each descriptor set.  The
  inner list contains model accuracy measures for each model fit
  to the data.  Regression models are assessed with Pearson's \eqn{R^2} and
  \eqn{RMSE}. Classification models are assessed with contingency tables.}
\item{classify}{a logical.  Were classification models used for binary
  response?}
\item{responses}{a numeric vector.  The true value of the response.}
\item{data}{a list of numeric matrices.  Each matrix is a descriptor set used
  as model input. The first column of each matrix is the response vector, and
  the remaining columns are descriptors.}
\item{params}{a list of data frames as made by
  \code{\link{MakeModelDefaults}}.  Each data frame contains the parameters to
  be set for a particular model.}
}
\description{
\code{ModelTrain} fits a series of classification or regression
models to sets of descriptors and computes cross-validated measures
of model performance.
}
\details{
\code{ModelTrain} fits a number of model fitting operations for each
data set provided. The response variable should either be binary or
continuous.  Descriptors can be any object as long as the models used
can take the object as input.

Not all modeling strategies will be appropriate for all response
types. For example, partial least squares linear discriminant analysis
("PLSLDA")
is not directly appropriate for continuous response assays such as
percent inhibition, but it can be applied once a threshold value for
percent inhibition is used to create a binary (active/inactive) response.

See \url{https://jrash.github.io/} for more information about the
models available (including model default parameters).

Sensible default values are selected for each
tunable model parameter, however users may set any parameter
manually using \code{\link{MakeModelDefaults}} and \code{user.params}.

\code{ModelTrain} predictions are based on k-fold cross-validation,
where the dataset is randomly divided into k parts, each containing
approximately equal numbers of compounds. Treating one of these parts
as a "test set" the remaining
k-1 parts are combined together as a "training set"
and used to build a model from the desired modeling technique and
descriptor set. This model is then applied to the "test set" to obtain
predictions. The process is repeated, holding out each of the k parts
in turn. One advantage of k-fold cross-validation is reduction in bias
from using the same data to both build and assess a model. Another
advantage is the increased precision of error estimation offered by
k-fold cross validation over a one-time split.

Recognizing that the definition of folds in k-fold cross validation
may have an impact on the observed performance measures, all models
are built using the same definition of folds. This process is repeated
to obtain multiple separate k-fold cross validation runs resulting in
multiple separate definitions of folds.  The number of these "splits"
is specified by \code{nsplits}.

Multiple descriptor sets can be specified
by the user. For each descriptor set, repeated k-fold cross validation
is performed for the spcified number of regression and/or classification
models.
}
\examples{
cml <- ModelTrain(USArrests, nsplits = 3)
cml

bin <- rbinom(50, 1, .1)
cml <- ModelTrain(cbind(bin, USArrests), nsplits = 3)
cml
}
\author{
Jacqueline Hughes-Oliver, Jeremy Ash
}
\references{
?
}
\seealso{
\code{\link{chemmodlab}}, \code{\link{plot.chemmodlab}},
  \code{\link{CombineSplits}},
}

