% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot.chemmodlab}
\alias{plot.chemmodlab}
\title{Plot method for the chemmodlab class.}
\usage{
\method{plot}{chemmodlab}(cml.result, max.select = NA)
}
\arguments{
\item{cml.result}{an object of class \code{\link{chemmodlab}}.}

\item{max.select}{the maximum number of tests to plot for the
accumulation curve. If max.select not specified, use \code{floor(min(300,n/4))},
where n is the number of compounds.}
}
\description{
\code{plot.chemmodlab} takes a \code{\link{chemmodlab}} object output by the
\code{\link{ModelTrain}} function and creates a series of accumulation curve
plots for assesing model and descriptor set performance.
}
\details{
For a binary response, the accumulation curve plots the number of assay hits
identified as a function of the number of tests conducted, where testing
order is determined by the predicted probabilities obtained from k-fold cross
validation. Given a particular compound collection, larger accumulations are
preferable.

The accumulation curve has also been extended to continuous responses.
Assuming large values of a continuous response y are preferable, ChemModLab
accumulates y so that \eqn{\sum_{i=1}^n y_i}{\sum y_i} is the sum of the y
over the first n tests. This extension includes the binary-response
accumulation curve as a special case.

By default, we display accumulation curves up to 300 tests, not for the
entire collection, to focus on the goal of finding actives as early as
possible.

There are two main series of plots generated:
}
\section{First plot series}{

 There is one plot per split and descriptor set
 combination. The accumulation curve for each model is compared.
}

\section{Second plot series}{

 There is one plot per split and model fit. The
 accumulation curve for each descriptor set combination is
 compared.
}
\examples{
cml <- modelTrain(USArrests, nsplits = 3)
plot(cml)

bin <- rbinom(50, 1, .1)
cml <- ModelTrain(cbind(bin, USArrests), nsplits = 3)
plot(cml)

}
\author{
Jacqueline Hughes-Oliver, Jeremy Ash
}
\references{
Modified from code originally written by
  William J. Welch 2001-2002
}
\seealso{
\code{\link{chemmodlab}}, \code{\link{modelTrain}}, \code{\link{modelTrain}}
}

